requirements.txt
google-genai>=0.5.0
websockets>=12.0
asyncio 